Renamed _property_serialization.py to _types.py
