+++ webviz_subsurface/plugins/_group_tree/_ensemble_group_tree_data.py
@@ -7,7 +7,7 @@ from webviz_config.common_cache import CACHE
 
 from webviz_subsurface._providers import EnsembleSummaryProvider
 
-from ._property_serialization import StatOptions, TreeModeOptions
+from ._types import StatOptions, TreeModeOptions
 
 
 class EnsembleGroupTreeData:
